requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
ollama>=0.1.6
firecrawl-py>=0.0.5
llama-index>=0.9.14
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-huggingface>=0.1.3
llama-index-vector-stores-chroma>=0.1.4
chromadb>=0.4.18
sentence-transformers>=2.2.0,<3.0.0
gradio>=4.7.1
celery[redis]>=5.3.4
redis>=5.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
requests>=2.31.0
aiofiles>=23.2.1
pytest>=7.4.3
pytest-cov>=4.1.0
locust>=2.17.0
numpy>=1.17.3,<1.25.0
black>=25.0.0
flake8>=7.0.0
mypy>=1.0.0
transformers>=4.34.0,<4.40.0